Nexus in the middle  (vk/50be-nexus-in-the-mid)
Add the hex view of the main planet to the save state.

div in App at /src/App.jsx:139

    └─ div:nth-child(2) in App at /src/App.jsx:330

        └─ div:nth-child(3) in App at /src/App.jsx:375

            └─ div in App at /src/App.jsx:377

                └─ svg at /src/components/game/HexGrid.jsx:148

Make sure that the very middle (0, 0) tile is always built and it's the Nexus building. When pressing a hex like this
div in App at /src/App.jsx:139

    └─ div:nth-child(2) in App at /src/App.jsx:330

        └─ div:nth-child(3) in App at /src/App.jsx:375

            └─ div in App at /src/App.jsx:377

                └─ svg at /src/components/game/HexGrid.jsx:148

                    └─ g at /src/components/game/HexGrid.jsx:153

                        └─ g:nth-child(67) at /src/components/game/HexGrid.jsx:163

                            └─ polygon at /src/components/game/HexGrid.jsx:169

display a building menu that let's you build a building of any category. Building takes 10 seconds.
When pressing on an already existing building add an option to demolish it.