Add a new building type (vk/2b8a-add-a-new-buildi)
Add a Logistics Center building using (HANGAR\_BAY) icon. When constructed the system acts as a link between all connected suppliers and consumers.
If A is a producer, and C is a consumer currently the only way to satisfy the demand is to construct the FTL:
A -> C
If a system B has a Logistics Center
A -> B -> C
would also satisfy the demand and the FTLs between A and B, as well as B and C pass the supply to the demanding system. FTL Tether animations should also reflect the flow of the resources